Update how we handle sorting
Now I'm sorting the list of all open PRs using the api.
Also, I'm ensuring that the UI is also sorting by created. This is the
best way to ensure that "trees" of PRs are represented in full.
although, it is entirely possible for someone to create PR 2 first, only
to later create another PR that goes before it. And the UI doesn't allow
for per_page so the tree is broken across pages.